Extract worker with modal dialog as a class.
